Move client code to client.rs
